+++ timelock-agent/src/main/java/com/palantir/timelock/config/TimeLockRuntimeConfiguration.java
@@ -4,15 +4,12 @@
 
 package com.palantir.timelock.config;
 
-import java.util.Set;
-
 import org.immutables.value.Value;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
 import com.fasterxml.jackson.databind.annotation.JsonSerialize;
 import com.google.common.base.Preconditions;
-import com.palantir.atlasdb.timelock.paxos.PaxosTimeLockConstants;
 
 /**
  * Dynamic (live-reloaded) portions of TimeLock's configuration.
@@ -21,14 +18,21 @@ import com.palantir.atlasdb.timelock.paxos.PaxosTimeLockConstants;
 @JsonSerialize(as = ImmutableTimeLockRuntimeConfiguration.class)
 @Value.Immutable
 public abstract class TimeLockRuntimeConfiguration {
-    private static final String CLIENT_NAME_REGEX = "[a-zA-Z0-9_-]+";
 
     @Value.Default
     public PaxosRuntimeConfiguration paxos() {
         return ImmutablePaxosRuntimeConfiguration.builder().build();
     }
 
-    public abstract Set<String> clients();
+    /**
+     * The maximum number of client namespaces to allow. Each distinct client consumes some amount of memory and disk
+     * space.
+     */
+    @JsonProperty("max-number-of-clients")
+    @Value.Default
+    public Integer maxNumberOfClients() {
+        return 100;
+    }
 
     /**
      * Log at INFO if a lock request receives a response after given duration in milliseconds.
@@ -44,12 +48,5 @@ public abstract class TimeLockRuntimeConfiguration {
     public void check() {
         Preconditions.checkState(slowLockLogTriggerMillis() >= 0,
                 "Slow lock log trigger threshold must be nonnegative, but found %s", slowLockLogTriggerMillis());
-        clients().forEach(client -> Preconditions.checkState(
-                client.matches(CLIENT_NAME_REGEX),
-                "Client names must consist of alphanumeric characters, underscores, or dashes. Illegal name: %s",
-                client));
-        Preconditions.checkState(!clients().contains(PaxosTimeLockConstants.LEADER_ELECTION_NAMESPACE),
-                "The client name '%s' is reserved for the leader election service, and may not be used.",
-                PaxosTimeLockConstants.LEADER_ELECTION_NAMESPACE);
     }
 }